requirements.txt
Flask==1.1.2
folium==0.11.0
geopandas==0.8.1
gunicorn==20.1.0
numpy==1.20.2
pandas==1.2.4
PyYAML==5.4.1
requests==2.25.1
Rtree==0.9.7
Shapely==1.7.1
joblib==1.0.1
